Reorganize tiny-orm-model
